ryan/code/test/requirements.txt
Pillow==10.0.1
Pillow==10.4.0
Requests==2.32.3
transformers==4.42.4
transformers==4.32.1
